Update IFBeamDBInterfaceV2.cpp
When the tool fetches information and writes to the TTree, if a device was empty/didn't return any values from the database, the entire branch will not be written to the root file to ensure the number of entries is the same between all branches. If there are no entries for a fetched device, fill the "blank" entries with -9999.