Update for loading LAPPD from ANNIEEvent booststore
While loading the event from processed data, the tool need to load the
run number from the boost store to index the ID config, update that.

Also, if the loaded LAPPD_ID is larger than 20, it will be treated as
the Incom ID, because I don't think we will have more than 20 LAPPDs at
the same time, and the earliest LAPPD we ever seen is 25 (or 40).

Add the default pedestal path in the LAPPDProcessedAna config, also add
the LAPPD ID config file name.